+++ code/game/objects/items/inspector.dm
@@ -1,6 +1,3 @@
-///Energy used to say an error message.
-#define ENERGY_TO_SPEAK (0.001 * STANDARD_CELL_CHARGE)
-
 /**
  * # N-spect scanner
  *
@@ -29,8 +26,10 @@
 	var/obj/item/stock_parts/cell/cell = /obj/item/stock_parts/cell/crap
 	///Cell cover status
 	var/cell_cover_open = FALSE
-	///Energy used per print.
-	var/energy_per_print = INSPECTOR_ENERGY_USAGE_NORMAL
+	///Power used per print in cell units
+	var/power_per_print = INSPECTOR_POWER_USAGE_NORMAL
+	///Power used to say an error message
+	var/power_to_speak = 1
 
 /obj/item/inspector/Initialize(mapload)
 	. = ..()
@@ -112,8 +111,8 @@
 	if(cell.charge == 0)
 		to_chat(user, "<span class='info'>\The [src] doesn't seem to be on... Perhaps it ran out of power?")
 		return
-	if(!cell.use(energy_per_print))
-		if(cell.use(ENERGY_TO_SPEAK))
+	if(!cell.use(power_per_print))
+		if(cell.use(power_to_speak))
 			say("ERROR! POWER CELL CHARGE LEVEL TOO LOW TO PRINT REPORT!")
 		return
 
@@ -239,8 +238,8 @@
  *
  * Can print things way faster, at full power the reports printed by this will destroy
  * themselves and leave water behind when folding is attempted by someone who isn't an
- * origami master. Printing at full power costs INSPECTOR_ENERGY_USAGE_HONK cell units
- * instead of INSPECTOR_ENERGY_USAGE_NORMAL cell units.
+ * origami master. Printing at full power costs INSPECTOR_POWER_USAGE_HONK cell units
+ * instead of INSPECTOR_POWER_USAGE_NORMAL cell units.
  */
 /obj/item/inspector/clown/bananium
 	name = "\improper Bananium HONK-spect scanner"
@@ -259,11 +258,11 @@
 
 /obj/item/inspector/clown/bananium/proc/check_settings_legality()
 	if(print_sound_mode == INSPECTOR_PRINT_SOUND_MODE_NORMAL && time_mode == INSPECTOR_TIME_MODE_HONK)
-		if(cell.use(ENERGY_TO_SPEAK))
+		if(cell.use(power_to_speak))
 			say("Setting combination forbidden by Geneva convention revision CCXXIII selected, reverting to defaults")
 		time_mode = INSPECTOR_TIME_MODE_SLOW
 		print_sound_mode = INSPECTOR_PRINT_SOUND_MODE_NORMAL
-		energy_per_print = INSPECTOR_ENERGY_USAGE_NORMAL
+		power_per_print = INSPECTOR_POWER_USAGE_NORMAL
 
 /obj/item/inspector/clown/bananium/screwdriver_act(mob/living/user, obj/item/tool)
 	. = ..()
@@ -297,7 +296,7 @@
 	if(time_mode != INSPECTOR_TIME_MODE_HONK)
 		return ..()
 	if(paper_charges == 0)
-		if(cell.use(ENERGY_TO_SPEAK))
+		if(cell.use(power_to_speak))
 			say("ERROR! OUT OF PAPER! MAXIMUM PRINTING SPEED UNAVAIBLE! SWITCH TO A SLOWER SPEED TO OR PROVIDE PAPER!")
 		else
 			to_chat(user, "<span class='info'>\The [src] doesn't seem to be on... Perhaps it ran out of power?")
@@ -309,7 +308,7 @@
 	var/message
 	switch(time_mode)
 		if(INSPECTOR_TIME_MODE_HONK)
-			energy_per_print = INSPECTOR_ENERGY_USAGE_NORMAL
+			power_per_print = INSPECTOR_POWER_USAGE_NORMAL
 			time_mode = INSPECTOR_TIME_MODE_SLOW
 			message = "SLOW."
 		if(INSPECTOR_TIME_MODE_SLOW)
@@ -317,7 +316,7 @@
 			message = "LIGHTNING FAST."
 		else
 			time_mode = INSPECTOR_TIME_MODE_HONK
-			energy_per_print = INSPECTOR_ENERGY_USAGE_HONK
+			power_per_print = INSPECTOR_POWER_USAGE_HONK
 			message = "HONK!"
 	balloon_alert(user, "scanning speed set to [message]")
 
@@ -373,17 +372,15 @@
  */
 /obj/item/paper/fake_report/water
 	grind_results = list(/datum/reagent/water = 5)
-	interaction_flags_click = NEED_DEXTERITY|NEED_HANDS|ALLOW_RESTING
 
-/obj/item/paper/fake_report/water/click_alt(mob/living/user)
+/obj/item/paper/fake_report/water/AltClick(mob/living/user, obj/item/I)
+	if(!user.can_perform_action(src, NEED_DEXTERITY|NEED_HANDS))
+		return
 	var/datum/action/innate/origami/origami_action = locate() in user.actions
 	if(origami_action?.active) //Origami masters can fold water
-		make_plane(user, /obj/item/paperplane/syndicate)
+		make_plane(user, I, /obj/item/paperplane/syndicate)
 	else if(do_after(user, 1 SECONDS, target = src, progress=TRUE))
 		var/turf/open/target = get_turf(src)
 		target.MakeSlippery(TURF_WET_WATER, min_wet_time = 10 SECONDS, wet_time_to_add = 5 SECONDS)
 		to_chat(user, span_notice("As you try to fold [src] into the shape of a plane, it disintegrates into water!"))
 		qdel(src)
-	return CLICK_ACTION_SUCCESS
-
-#undef ENERGY_TO_SPEAK
